// src/types.ts
export interface SiteSettings {
  brand:                string;
  phone:                string;
  email:                string;
  currency:             string;
  whatsapp:             string;
  telegram:             string;
  telegram_bot_token?:  string;
  telegram_chat_id?:    string;
  instapay_username?:   string;
  vodafone_cash?:       string;
  bank_details?:        string;
  payment_instructions?: string;
}